More proper CSS and added blinking functionality to the text-cursor. Also various code clean-ups such as naming and removing unnecessary CSS.
Previously, text that extends beyond a certain point would cause the div
for the command input to be shifted one line down from the prompt. We
want the prompt and command to be connected.

Furthermore, along the point of longer text, the previous CSS would
overflow the screen so that scrolling is enabled. Instead, I made it
wrap back to the start of the line.

There were a number of other weird CSS "bugs" regarding size of lines
and margins. They have been mostly improved.